pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "goggles-ai"
version = "0.1.0"
description = "goggles-ai — open-source inspection layer for AI agent inputs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
deep = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]
test = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "responses>=0.25.0",
]
dev = [
    "goggles-ai[test]",
    "black>=24.3.0",
    "ruff>=0.4.0",
]
all = [
    "goggles-ai[deep,test,dev]",
]

[project.scripts]
goggles-ai = "goggles_ai.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["goggles_ai"]
omit = ["goggles_ai/middleware/*", "goggles_ai/detectors/stego_deep.py"]

[tool.hatch.build.targets.wheel]
packages = ["goggles_ai"]
